requirements.txt
streamlit
openai==0.28.0
python-dotenv
python-docx
PyMuPDF
